more Array things...
My rationale for the "new" desugaring change is to prevent interpreter
exceptions from bubbling up to the top.  If the argument to new doesn't
define "prototype", it leads to an object that violates get_field's
assumption about the #proto chain (i.e. that all #proto's are either
objects or null).