More cleanups
Add mkArgsObj to the environment to reduce clutter in desugared code.

Also, don't repeat the function expression so many times.

We need to be really careful in desugaring that each expression in the
original program only gets evaluated once --- this is another place
where side effects in an expression would be duplicated by subtle
desugaring problems.

Finally, "typeof" was doing what we wanted in appexpr_check, anyway.
It reports "function" for objects that have a code attribute.  "lambda"
is returned for bare closures, but this should never be exposed to the
JavaScript program.