Getting eval to work for CPS
Added EGetFieldK and ESetFieldK to the interpreter.

Add extra arguments to get_field and set_field:

 - getter_args/setter_args --- new arguments to pass on to getter/setter
 - result --- Ocaml function to call with the result, if a normal lookup

Change de_cps to use these forms.

Futz around a little bit with testing: cps versions of the shell scripts, and
add a new env, cps.env, that is wrapped around things to provide %answer and
%error.