Fix defineProperty, and add a few more (not enough yet) tests
defineProperty had a free variable called current-prop floating around.

This was meant to be a property descriptor describing the property to be
updated, but it was never assigned to anything.  The indentation and the
nesting of th ifs in the function were also badly broken.

Test cases for some known corner cases were added, but more would be nice.