SymObj keeps track of locs for making sym vals for unknown fields.
When getting an unknown field from a sym obj, we used to just allow it
to be a pointer to any obj in the store. Really it should only be able
to point to objs that were in the store when the sym obj was created,
just like we do with prototypes.