Replaced env with env stack, abstracting away the representation.
This will allow the garbage collector to walk the entire stack when
determining its root set (i.e. it will be able to find shadowed
bindings). This stack representation is now abstracted away so that
the rest of the evaluator can continue to behave as if it only has one
flat env.