Finished sym GC. Replaced env with env stack.
The env stack represents all the environments on the call stack at any
given time. The GC needs to look through all of them to find the root
set. For now, each env is just a list of id-loc pairs, which is less
efficient for lookup but better for the GC. We can optimize this
later.