Some really subtle and awkward logic for globals/eval
Globals and eval still weren't quite happy with one another, so
defineGlobalAccessor got more logic for determining when a free variable in an
eval needs to be defined with a getter that throws a ReferenceError.
defineGlobalAccessor still needs more review; it's not perfect yet

Fixed a duplicated code but in the desugaring of delete, introducing a let
binding.

Fixed free variables checking for forin/forvar statements, which weren't
getting appropriately removed, and were being added as global accessors.

Moved the declaration of arguments the variable into the context object, to
avoid going up the scope chain if arguments appears gets added as an accessor
to %globalContext (more review should determine if this should even be
possible).

This commit makes test_GLOBAL_LEAKS_FROM_GLOBAL_FUNCTION_CALLS and
test_GLOBAL_LEAKS_FROM_GLOBALLY_CALLED_BUILTINS pass, which are now unit tests
ses-global-leak{1,2}.js