More fiddling with execution contexts.
The comment in eval was right --- I've switched strict off of contexts' primval
and into a lexical binding of the S5 variable #strict.  This makes property
lookup, and probably other future features, more annoying.  However, having it
on the global context was just wrong because of indirect, non-strict evals.

A bunch of tests from chapter 10 help support the change