Change eval() to take an environment object as an argument
Surprisingly few changes here.

The upshot is that @eval() now takes two arguments, a string and an
object whose bindings will be the environment of the desugared string.
This is really akin to E and Racket's eval.

Added a function, %makeGlobalEnv, that produces an object with a field
for every global binding.  We might want to think about automating its
creation, since you'll have to remember every time you add a binding to
es5.env.  It also might be possible to get by with a smaller subset,
depending on what desugared code actually has as free variables (this is
a superset of those).

I used this grep to find the identifiers needed for makeGlobal:

grep "^let \[%[-_a-zA-Z0-9]*\]" envs/es5.env

And some vim regexps to take the output of that and turn it into
assignments.  Not actually that painful.  All expected (unit) tests
pass.  CPS got stubbed-out changes, but Eval is sort of broken there
anyway.