Consolidate all free env variables used in desugaring
The new module env_free_vars contains all the variables used by
desugaring.  When a variable from the environment is used in desugaring,
it can call F.env_var to check that the variable is indeed expected.

Env_free_vars also defines the %makeGlobalEnv function, which can be
scoped to only those free variables that the desugared code actually
needs (plus bindings like %nonstrictContext and %this added by eval()).

To accomplish this, a new command line option was added, -internal-env,
which takes a named internal environment (defined by the main function),
and loads it directly.

I also cleaned up some of the directly named variables in exprjs_to_ljs,
but this is a bigger, separate TODO.