Significant refactoring to exceptions re: discussion with blerner
There has been longstanding ickiness with the way exceptions are handled
in S5, especially when signalling errors from within the interpreter.
The problem was that S5 code couldn't distinguish between errors from
JavaScript and errors from the interpreter.  This made it difficult to
e.g. signal an actual TypeError on an unwritable field assignment
without checking for writableness in desugaring first.  A series of
kludges work to circumvent this, but are largely dissatisfying.

To remedy this, blerner suggested desugaring JavaScript throw
expressions to put the actual thrown value one layer deep in a property.
Catches are then desugared to look for this property, which I've named
"%js-exn", and can handle non-JS error specially, while handing JS
errors onto the desugared catch block when appropriate.

This refactoring involved changing every throw statement in the
environment that throws an exception for JavaScript's observation, and
tweaking a few desugaring rules.  The main benefit of refactoring is in
the new %DispatchError function in the environment, which can now serve
as an authoritative mapping of interpreter errors to JavaScript errors.
This can now be called from the several places that check for writable
field mishaps, and can be extended to address field deletion, field
extension, etc, which are currently handled by verbose desugaring.

This adds %IsJSError and %ErrorDispatch to the free variables of
desugaring.  %IsJSError is merely a predicate, and %ErrorDispatch
*always* throws, and either rethrows the value that came in, or creates
and throws one of several built-in errors.